# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinkedIn GPT API",
    "description": "API for accessing LinkedIn data, including top voices, search, profiles, and authentication.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://db828f2e-5d46-40d2-9a43-3f782fc3f1d8-00-2nhlwjy3wyleq.sisko.replit.dev/api/v1"
    }
  ],
  "paths": {
...
    "/api/v1/profiles/analyze": {
      "post": {
        "description": "Analyze LinkedIn profile",
        "operationId": "analyzeProfiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profileUrl": {
                    "type": "string",
                    "items": {
                      "type": "string"
                    },
                    "description": "LinkedIn profile URL to analyze"
                  },
                  "subscriptionId": {
                    "type": "string",
                    "description": "Subscription ID for premium features"
                  }
                },
                "required": ["profileUrls", "subscriptionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile analysis "
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Subscription required"
          }
        },
        "tags": ["Profile"]
      }
    },
    "/api/v1/profiles/posts/{profileId}": {
      "get": {
        "description": "Get posts from a specific LinkedIn profile",
        "operationId": "getProfilePosts",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "description": "LinkedIn profile ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile posts"
          },
          "403": {
            "description": "Subscription required"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "tags": ["Profiles"]
      }
    },
    "/api/v1/profiles/status/{requestId}": {
      "get": {
        "description": "Check status of a profile analysis request",
        "operationId": "getAnalysisStatus",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "description": "Request ID from analyze call",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID for verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analysis status"
          },
          "403": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "tags": ["Profiles"]
      }
    },
    "/api/v1/profiles/recent": {
      "get": {
        "description": "Get recently analyzed profiles",
        "operationId": "getRecentProfiles",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent profiles"
          },
          "403": {
            "description": "Subscription required"
          }
        },
        "tags": ["Profiles"]
      }
    },
    "/api/v1/profiles/{profileId}": {
      "delete": {
        "description": "Delete a profile analysis",
        "operationId": "deleteProfileAnalysis",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "description": "Profile ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile analysis deleted"
          },
          "403": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "tags": ["Profiles"]
      }
    },
...
  },
...
}